+++ R/spinners.R
@@ -16,18 +16,18 @@
 #'
 #' @examples spin_rotating_plane()
 #' 
+#' @return An object of class \code{spinner}.
+#' 
 #' @name spinners
 #' @export
 spin_rotating_plane <- function(){
-  kit(1)
-  invisible(div(class = "rotating-plane"))
+  construct(div(class = "rotating-plane"), 1)
 }
 
 #' @rdname spinners
 #' @export
 spin_fading_circles <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "sk-fading-circle",
       div(class = "sk-circle1 sk-circle"),
@@ -42,43 +42,43 @@ spin_fading_circles <- function(){
       div(class = "sk-circle10 sk-circle"),
       div(class = "sk-circle11 sk-circle"),
       div(class = "sk-circle12 sk-circle")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_folding_cube <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "sk-folding-cube",
       div(class = "sk-cube1 sk-cube"),
       div(class = "sk-cube2 sk-cube"),
       div(class = "sk-cube3 sk-cube"),
       div(class = "sk-cube4 sk-cube")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_double_bounce <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "double-bouncer",
       div(class = "double-bounce1"),
       div(class = "double-bounce2")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_wave <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "waver",
       div(class = "rect1"),
@@ -86,62 +86,61 @@ spin_wave <- function(){
       div(class = "rect3"),
       div(class = "rect4"),
       div(class = "rect5")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_wandering_cubes <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "wandering-cubes",
       div(class = "cube1"),
       div(class = "cube2")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_pulse <- function(){
-  kit(1)
-  invisible(div(class = "pulser"))
+  construct(div(class = "pulser"), 1)
 }
 
 #' @rdname spinners
 #' @export
 spin_chasing_dots <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "chasing-dots",
       div(class = "dot1"),
       div(class = "dot2")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_three_bounce <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "three-bounce",
       div(class = "bounce1"),
       div(class = "bounce2"),
       div(class = "bounce3")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_circle <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "sk-circle",
       div(class = "sk-circle1 sk-child"),
@@ -156,30 +155,30 @@ spin_circle <- function(){
       div(class = "sk-circle10 sk-child"),
       div(class = "sk-circle11 sk-child"),
       div(class = "sk-circle12 sk-child")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_rotate <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "spinner-box",
       div(
         class = "circle-border",
         div(class = "circle-core")
       )
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_solar <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "spinner-box",
       div(
@@ -201,15 +200,15 @@ spin_solar <- function(){
           )
         )
       )
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_orbit <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "spinner-box",
       div(class = "blue-orbit leo"),
@@ -218,15 +217,15 @@ spin_orbit <- function(){
       div(class = "white-orbit w1 leo"),
       div(class = "white-orbit w2 leo"),
       div(class = "white-orbit w3 leo")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_squares <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "spinner-box",
       div(
@@ -237,15 +236,15 @@ spin_squares <- function(){
         class = "configure-border-2",
         div(class = "configure-core")
       )
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_cube_grid <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "sk-cube-grid",
       div(class = "sk-cube sk-cube1"),
@@ -257,22 +256,21 @@ spin_cube_grid <- function(){
       div(class = "sk-cube sk-cube7"),
       div(class = "sk-cube sk-cube8"),
       div(class = "sk-cube sk-cube9")
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_circles <- function(){
-  kit(2)
-	invisible(div(class = "lds-circle"))
+	construct(div(class = "lds-circle"), 1)
 }
 
 #' @rdname spinners
 #' @export
 spin_orbiter <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "orbiter-spinner",
       div(
@@ -280,22 +278,21 @@ spin_orbiter <- function(){
         class = "orbiter",
         class = "orbiter"
       )
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_pixel <- function(){
-  kit(1)
-	invisible(div(class = "pixel-spinner", div(class = "pixel-spinner-inner")))
+	construct(div(class = "pixel-spinner", div(class = "pixel-spinner-inner")), 1)
 }
 
 #' @rdname spinners
 #' @export
 spin_flower <- function(){
-  kit(1)
-  invisible(
+  construct(
     div(
       class = "flower-spinner",
       div(
@@ -307,240 +304,216 @@ spin_flower <- function(){
           )
         )
       )
-    )
+    ),
+    1
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_dual_ring <- function(){
-  kit(2)
-	invisible(div(class = "lds-dual-ring"))
+	construct(div(class = "lds-dual-ring"), 2)
 }
 
 #' @rdname spinners
 #' @export
 spin_heart <- function(){
-  kit(2)
-  invisible(div(class = "lds-heart", div()))
+  construct(div(class = "lds-heart", div()), 2)
 }
 
 #' @rdname spinners
 #' @export
 spin_ellipsis <- function(){
-  kit(2)
-  invisible(
+  construct(
     div(
       class = "lds-ellipsis",
       div(),
       div(),
       div(),
       div()
-    )
+    ),
+    2
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_facebook <- function(){
-  kit(2)
-  invisible(
+  construct(
     div(
       class = "lds-facebook",
       div(),
       div(),
       div()
-    )
+    ),
+    2
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_hourglass <- function(){
-  kit(2)
-	invisible(div(class = "lds-hourglass"))
+	construct(div(class = "lds-hourglass"), 2)
 }
 
 #' @rdname spinners
 #' @export
 spin_ring <- function(){
-  kit(2)
-  invisible(
+  construct(
     div(
       class = "lds-ring",
       div(),
       div(),
       div(),
       div()
-    )
+    ),
+    2
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_ripple <- function(){
-  kit(2)
-  invisible(
+  construct(
     div(
       class = "lds-ripple",
       div(),
       div()
-    )
+    ),
+    2
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_terminal <- function(){
-  kit(2)
-  invisible(div(class = "waiter-terminal"))
+  construct(div(class = "waiter-terminal"), 2)
 }
 
 #' @rdname spinners
 #' @export
 spin_loader <- function(){
-  kit(4)
-  invisible(div(class = "spinner-loader"))
+  construct(div(class = "spinner-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_throbber <- function(){
-  kit(4)
-  invisible(div(class = "throbber-loader"))
+  construct(div(class = "throbber-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_refresh <- function(){
-  kit(4)
-  invisible(div(class = "refreshing-loader"))
+  construct(div(class = "refreshing-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_heartbeat <- function(){
-  kit(4)
-  invisible(div(class = "heartbeat-loader"))
+  construct(div(class = "heartbeat-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_gauge <- function(){
-  kit(4)
-  invisible(div(class = "gauge-loader"))
+  construct(div(class = "gauge-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_3k <- function(){
-  kit(4)
-  invisible(div(class = "three-quarters-loader"))
+  construct(div(class = "three-quarters-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_wobblebar <- function(){
-  kit(4)
-  invisible(div(class = "wobblebar-loader"))
+  construct(div(class = "wobblebar-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_atebits <- function(){
-  kit(4)
-  invisible(div(class = "atebits-loader"))
+  construct(div(class = "atebits-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_whirly <- function(){
-  kit(4)
-  invisible(div(class = "whirly-loader"))
+  construct(div(class = "whirly-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_flowers <- function(){
-  kit(4)
-  invisible(div(class = "flower-loader"))
+  construct(div(class = "flower-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_dots <- function(){
-  kit(4)
-  invisible(div(class = "dots-loader"))
+  construct(div(class = "dots-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_3circles <- function(){
-  kit(4)
-  invisible(div(class = "circles-loader"))
+  construct(div(class = "circles-loader"))
 }
 
 #' @rdname spinners
 #' @export
 spin_plus <- function(){
-  kit(4)
-  invisible(div(class = "plus-loader"))
+  construct(div(class = "plus-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_pulsar <- function(){
-  kit(4)
-  invisible(div(class = "pulse-loader"))
+  construct(div(class = "pulse-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_hexdots <- function(){
-  kit(4)
-  invisible(div(class = "hexdots-loader"))
+  construct(div(class = "hexdots-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_inner_circles <- function(){
-  kit(4)
-  invisible(div(class = "inner-circles-loader"))
+  construct(div(class = "inner-circles-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_pong <- function(){
-  kit(4)
-  invisible(div(class = "pong-loader"))
+  construct(div(class = "pong-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_timer <- function(){
-  kit(4)
-  invisible(div(class = "timer-loader"))
+  construct(div(class = "timer-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_ball <- function(){
-  kit(4)
-  invisible(div(class = "ball-loader"))
+  construct(div(class = "ball-loader"), 4)
 }
 
 #' @rdname spinners
 #' @export
 spin_dual_circle <- function(){
-  kit(5)
-  invisible(div(class = "sbl-circ-dual"))
+  construct(div(class = "sbl-circ-dual"), 5)
 }
 
 #' @rdname spinners
 #' @export
 spin_seven_circle <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-seven-circles",
       div(),
@@ -550,85 +523,81 @@ spin_seven_circle <- function(){
       div(),
       div(),
       div()
-    )
+    ),
+    5
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_clock <- function(){
-  kit(5)
-  invisible(div(class = "sbl-meter"))
+  construct(div(class = "sbl-meter"), 5)
 }
 
 #' @rdname spinners
 #' @export
 spin_pushing_shapes <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-pushing-shapes",
       div(),
       div()
-    )
+    ),
+    5
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_fill <- function(){
-  kit(5)
-  invisible(div(class = "sbl-rect-spin-fill"))
+  construct(div(class = "sbl-rect-spin-fill"), 5)
 }
 
 #' @rdname spinners
 #' @export
 spin_rhombus <- function(){
-  kit(5)
-  invisible(div(class = "sbl-cirle-to-rhombus"))
+  construct(div(class = "sbl-cirle-to-rhombus"), 5)
 }
 
 #' @rdname spinners
 #' @export
 spin_balance <- function(){
-  kit(5)
-  invisible(div(class = "sbl-cirle-balance"))
+  construct(div(class = "sbl-cirle-balance"), 5)
 }
 
 #' @rdname spinners
 #' @export
 spin_square_circle <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-cirle-and-square",
       div(),
       div(),
       div(),
       div()
-    )
+    ),
+    5
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_circle_square <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-square-to-circle",
       div(),
       div(),
       div()
-    )
+    ),
+    5
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_puzzle <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-puzzle",
       div(),
@@ -636,19 +605,20 @@ spin_puzzle <- function(){
       div(),
       div(),
       div()
-    )
+    ),
+    5
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_half <- function(){
-  kit(5)
-  invisible(
+  construct(
     div(
       class = "sbl-half-circle-spin",
       div()
-    )
+    ),
+    5
   )
 }
 
@@ -666,15 +636,13 @@ spin_loaders <- function(id = 1, color = "white", style = NULL){
   else
     style <- color
   
-  kit(6)
-  invisible(div(class = paste0("loaderz-", id), style = style))
+  construct(div(class = paste0("loaderz-", id), style = style), 6)
 }
 
 #' @rdname spinners
 #' @export
 spin_1 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -683,15 +651,15 @@ spin_1 <- function(){
           class = "spinner spinner--1"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_2 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -700,15 +668,15 @@ spin_2 <- function(){
           class = "spinner spinner--2"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_3 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -717,15 +685,15 @@ spin_3 <- function(){
           class = "spinner spinner--3"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_4 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -734,15 +702,15 @@ spin_4 <- function(){
           class = "spinner spinner--4"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_5 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -751,15 +719,15 @@ spin_5 <- function(){
           class = "spinner spinner--5"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_6 <- function(){
-  kit(3)
-  invisible(
+  construct(
     div(
       class = "container--box",
       div(
@@ -768,15 +736,15 @@ spin_6 <- function(){
           class = "spinner spinner--6"
         )
       )
-    )
+    ),
+    3
   )
 }
 
 #' @rdname spinners
 #' @export
 spin_google <- function(){
-  kit(7)
-  invisible(div(class = "cm-spinner"))
+  construct(div(class = "cm-spinner"), 7)
 }
 
 #' @rdname spinners
@@ -784,5 +752,108 @@ spin_google <- function(){
 browse_waiters <- function() {
   .Deprecated("", package = "waiter", msg = "This function is no longer supported; see the package website.")
 	#shiny::runApp(appDir = system.file("waiter", package = 'waiter', mustWork = TRUE))
+  construct()
+}
+
+#' Preview spinner
+#' 
+#' Allows previewing spinners in web browser or RStudio Viewer.
+#' 
+#' @param spinner A waiter \code{link{spinner}}.
+#' @param bg_color Background color.
+#' 
+#' @examples 
+#' preview_spinner(spin_1())
+#' 
+#' @export
+preview_spinner <- function(spinner, bg_color = "black") UseMethod("preview_spinner")
+
+#'@export
+preview_spinner.spinner <- function(spinner, bg_color = "black"){
+
+  if(missing(spinner))
+    stop("Missing spinner", call. = FALSE)
+
+  deps <- get_spinner_dependencies()
+  spinner <- htmltools::attachDependencies(spinner, deps)
+
+  htmltools::browsable(
+    tagList(
+      tags$head(
+        tags$style(
+          paste0("body{background: ", bg_color,";}")
+        )
+      ),
+      div(
+        style = "position: relative;height:90vh;",
+        div(
+          style = "position: absolute;top:50%;left:50%;",
+          spinner
+        )
+      )
+    )
+  )
+
+}
+
+#' Get Dependencies
+#' 
+#' Get dependencies for preview.
+#' 
+#' @keywords internal
+get_spinner_dependencies <- function(){
+  htmltools::htmlDependency(
+    name = "css-spinner",
+    version = "1.0.0",
+    src = "assets/waiter",
+    package = "waiter",
+    stylesheet = c(
+      "css-spinners.css",
+      "devloop.css",
+      "please-wait.css",
+      "spinbolt.css",
+      "spinkit.css",
+      "spinners.css"
+    ),
+    all_files = TRUE
+  )
+}
+
+#' Spinner Constructor
+#' 
+#' Construct spinner class.
+#' 
+#' @param spinner Spinner object.
+#' @param id Id of spinkit to use.
+#' 
+#' @keywords internal
+construct <- function(spinner, id){
+  # hold spinner and id
+  attr(spinner, "id") <- id
+  structure(spinner, class = c("spinner", class(spinner)))
+}
+
+#' @export
+print.spinner <- function(x, ...){
+  kit(x)
+}
+
+#' Give Spinner kit id
+#' 
+#' Prints id of spinner CSS required.
+#' 
+#' @param spinner Spinner to print kit of.
+#' 
+#' @keywords internal
+kit <- function(spinner){
+  id <- attr(spinner, 'id')
+
+  cat(
+    "This spinner requires kit #",
+    crayon::cyan(id), ", include it with:\n", 
+    crayon::green("use_waiter(spinners = ", id, ")", sep = ""), " \n",
+    sep = ""
+  )
+
   invisible()
 }